refactor: Abstract to files
