+++ src/types.ts
@@ -1,0 +1,12 @@
+// Place some shared types here
+
+export interface ProjectInfo {
+  name: string;
+  email: string;
+  username: string;
+  repository: string;
+  proj_name: string;
+  proj_short_desc: string;
+  proj_long_desc: string;
+  docs_url: string;
+}
